+++ src/test/groovy/com/github/gradle/node/npm/task/NpxTaskTest.groovy
@@ -11,20 +11,19 @@ class NpxTaskTest extends AbstractTaskTest {
         execSpec = Mock(ExecSpec)
 
         def task = project.tasks.create('simple', NpxTask)
-        task.command = 'command'
-        task.args = ['a', 'b']
-        task.environment = ['a': '1']
-        task.ignoreExitValue = true
-        task.workingDir = projectDir
-        task.execOverrides = {}
+        task.command.set('command')
+        task.args.set(['a', 'b'])
+        task.environment.set(['a': '1'])
+        task.ignoreExitValue.set(true)
+        task.workingDir.set(projectDir)
 
         when:
         project.evaluate()
         task.exec()
 
         then:
-        task.command == 'command'
-        task.args == ['a', 'b']
+        task.command.get() == 'command'
+        task.args.get() == ['a', 'b']
         1 * execSpec.setIgnoreExitValue(true)
         1 * execSpec.setEnvironment({ it['a'] == '1' && containsPath(it) })
         1 * execSpec.setExecutable('npx')
@@ -38,20 +37,19 @@ class NpxTaskTest extends AbstractTaskTest {
         execSpec = Mock(ExecSpec)
 
         def task = project.tasks.create('simple', NpxTask)
-        task.command = 'command'
-        task.args = ['a', 'b']
-        task.environment = ['a': '1']
-        task.ignoreExitValue = true
-        task.workingDir = projectDir
-        task.execOverrides = {}
+        task.command.set('command')
+        task.args.set(['a', 'b'])
+        task.environment.set(['a': '1'])
+        task.ignoreExitValue.set(true)
+        task.workingDir.set(projectDir)
 
         when:
         project.evaluate()
         task.exec()
 
         then:
-        task.command == 'command'
-        task.args == ['a', 'b']
+        task.command.get() == 'command'
+        task.args.get() == ['a', 'b']
         1 * execSpec.setIgnoreExitValue(true)
         1 * execSpec.setEnvironment({ it['a'] == '1' && containsPath(it) })
         1 * execSpec.setExecutable('npx.cmd')
@@ -62,7 +60,7 @@ class NpxTaskTest extends AbstractTaskTest {
     def "exec npx task (download)"() {
         given:
         props.setProperty('os.name', 'Linux')
-        nodeExtension.download = true
+        nodeExtension.download.set(true)
         execSpec = Mock(ExecSpec)
         def variantComputer = new VariantComputer()
         def nodeDir = variantComputer.computeNodeDir(nodeExtension)
@@ -77,7 +75,7 @@ class NpxTaskTest extends AbstractTaskTest {
 
         then:
         1 * execSpec.setIgnoreExitValue(false)
-        1 * execSpec.setExecutable(new File(nodeBinDir, "node").toString())
-        1 * execSpec.setArgs([npxScriptFile])
+        1 * execSpec.setExecutable(nodeBinDir.get().file("node").asFile.toString())
+        1 * execSpec.setArgs([npxScriptFile.get()])
     }
 }